Improve build-book's format handling
Remove the ignored file feature in build-book in favor of format
detection.

In addition to using the presence of a ttm file to detect if a markdown
file is a literate Idris document or not, we also detect bare idris
files and feed them into a new katla wrapper that mangles the output of
katla's html backend into a suitable markdown document.